fix(signup): prevent customer creation b4 card verify
