+++ app/e2e/index.spec.ts
@@ -1,15 +1,7 @@
 import { expect, Page, test } from "@playwright/test";
-import axios from "axios";
-import crypto from "crypto";
-const BASE_URL = process.env.E2E_START_URL ?? "http://localhost:3000";
 
-// TODO: Share this context for Auth tests
-const EMAIL_DOMAINS_LIST = [];
-const SPONSOR_PLANS = ["$1 / Month", "$10 / Year"] as const;
-const EMAILS: Record<typeof SPONSOR_PLANS[number], string> = {
-  "$1 / Month": "",
-  "$10 / Year": "",
-};
+import { goToPath, goToTab } from "./utils";
+export const BASE_URL = process.env.E2E_START_URL ?? "http://localhost:3000";
 
 // Run in parallel
 test.describe.configure({ mode: "parallel" });
@@ -17,6 +9,7 @@ test.describe("Unauthorized", () => {
   test.beforeEach(async ({ page }) => {
     await goToPath(page);
   });
+
   test("Learn More -> Sponsors Page", async ({ page }) => {
     await goToTab(page, "Charts");
 
@@ -459,92 +452,7 @@ test.describe("Unauthorized", () => {
 
     expect(svg.suggestedFilename()).toBe("flowchart.svg");
   });
-
-  for (const plan of SPONSOR_PLANS) {
-    test.skip(`Sponsors > Become a ${plan} Sponsor`, async ({ page }) => {
-      test.setTimeout(60000);
-
-      const email = await getTempEmail();
-
-      // expect email not to be null
-      expect(email).toBeTruthy();
-
-      EMAILS[plan] = email;
-
-      await goToTab(page, "Sponsors");
-
-      // Click [data-testid="email"]
-      await page.locator('[data-testid="email"]').click();
-      // Fill [data-testid="email"]
-      await page.locator('[data-testid="email"]').fill(email);
-      // Click button[role="radio"]:has-text("$1 / Month")
-      await page.locator(`button[role="radio"]:has-text("${plan}")`).click();
-      // Click [data-testid="card-element"]
-      await page.locator('[data-testid="card-element"]').click();
-      // Click [placeholder="Card number"]
-      await page
-        .frameLocator('iframe[name*="__privateStripeFrame"]')
-        .locator('[placeholder="Card number"]')
-        .click();
-      // Fill [placeholder="Card number"]
-      await page
-        .frameLocator('iframe[name*="__privateStripeFrame"]')
-        .locator('[placeholder="Card number"]')
-        .fill("4242 4242 4242 4242");
-      // Fill [placeholder="MM \/ YY"]
-      await page
-        .frameLocator('iframe[name*="__privateStripeFrame"]')
-        .locator('[placeholder="MM \\/ YY"]')
-        .fill("01 / 30");
-      // Fill [placeholder="CVC"]
-      await page
-        .frameLocator('iframe[name*="__privateStripeFrame"]')
-        .locator('[placeholder="CVC"]')
-        .fill("222");
-      // Fill [placeholder="ZIP"]
-      await page
-        .frameLocator('iframe[name*="__privateStripeFrame"]')
-        .locator('[placeholder="ZIP"]')
-        .fill("22222");
-      // Click button:has-text("Sign Up")
-      await page.locator('button:has-text("Sign Up")').click();
-      // Click text=Check your email for a link to log in. You can close this window.
-      await expect(
-        page.locator(
-          "text=Check your email for a link to log in. You can close this window."
-        )
-      ).toBeVisible({ timeout: 60000 });
-
-      // Wait 5 seconds
-      await new Promise((resolve) => setTimeout(resolve, 5000));
-
-      // Make Sure Log in Email is sent
-      let emails = await getTempEmailMessage(EMAILS[plan]);
-
-      if (!emails.length) {
-        // Wait 5 seconds
-        await new Promise((resolve) => setTimeout(resolve, 5000));
-
-        emails = await getTempEmailMessage(EMAILS[plan]);
-      }
-
-      expect(emails.length).toBeGreaterThanOrEqual(1);
-      expect(
-        emails.some((email: { mail_html: string }) =>
-          /supabase.co\/auth\/v1\/verify/.test(email.mail_html)
-        )
-      ).toEqual(true);
-    });
-  }
 });
-
-async function goToPath(page: Page, path = "") {
-  await page.goto(`${BASE_URL}/${path}`);
-}
-
-async function goToTab(page: Page, tabName: string) {
-  await page.click(`button[role="tab"]:has-text("${tabName}")`);
-}
 
 async function changeEditorText(page: Page) {
   // Put cursor in editor
@@ -562,41 +470,4 @@ async function openExportDialog(page: Page) {
   page.locator('[aria-label="Open Export Dialog"]').click();
   // Click text=Download
   await expect(page.locator("text=Download")).toBeVisible({ timeout: 60000 });
-}
-
-/** Generates a temp email address */
-async function getTempEmail() {
-  if (!process.env.RAPID_API_KEY) throw new Error("Missing API Key");
-
-  if (EMAIL_DOMAINS_LIST.length === 0) {
-    const response = await axios({
-      method: "GET",
-      url: "https://privatix-temp-mail-v1.p.rapidapi.com/request/domains/",
-      headers: {
-        "X-RapidAPI-Host": "privatix-temp-mail-v1.p.rapidapi.com",
-        "X-RapidAPI-Key": process.env.RAPID_API_KEY,
-      },
-    });
-    EMAIL_DOMAINS_LIST.push(...response.data);
-  }
-
-  const randomDomain =
-    EMAIL_DOMAINS_LIST[Math.floor(Math.random() * EMAIL_DOMAINS_LIST.length)];
-  const randomEmail = `ci+${Date.now()}${randomDomain}`;
-  return randomEmail;
-}
-
-/** Returns inbox messages for a given email */
-async function getTempEmailMessage(email: string) {
-  const hash = crypto.createHash("md5").update(email).digest("hex");
-  const response = await axios.request({
-    method: "GET",
-    url: `https://privatix-temp-mail-v1.p.rapidapi.com/request/mail/id/${hash}/`,
-    headers: {
-      "X-RapidAPI-Host": "privatix-temp-mail-v1.p.rapidapi.com",
-      "X-RapidAPI-Key": process.env.RAPID_API_KEY,
-    },
-  });
-  const emails = response.data;
-  return emails;
 }
